+++ generators/python/src/fern_python/external_dependencies/pydantic.py
@@ -74,6 +74,10 @@ class Pydantic:
         def forbid(version: PydanticVersionCompatibility) -> AST.Expression:
             return AST.Expression(_export(version, "Extra", "forbid"))
 
+        @staticmethod
+        def allow(version: PydanticVersionCompatibility) -> AST.Expression:
+            return AST.Expression(_export(version, "Extra", "allow"))
+
     @staticmethod
     def root_validator(version: PydanticVersionCompatibility, pre: bool = False) -> AST.FunctionInvocation:
         return AST.FunctionInvocation(