+++ packages/cli/config-management/docs-configuration/src/index.ts
@@ -1,3 +1,4 @@
+export { getColorFromRawConfig, getColorType } from "./convertColorsConfiguration";
 export { getReferencedApiSections } from "./getReferencedApiSections";
 export * from "./ParsedDocsConfiguration";
 export { parseDocsConfiguration } from "./parseDocsConfiguration";