+++ packages/cli/ete-tests/src/tests/init/init.test.ts
@@ -46,11 +46,6 @@ describe("fern init", () => {
         expect(await getDirectoryContents(pathOfDirectory)).toMatchSnapshot();
     }, 60_000);
 
-    it("init openapi url", async () => {
-        const pathOfDirectory = await init({ openApiArg: "https://petstore3.swagger.io/api/v3/openapi.json" });
-        expect(await getDirectoryContents(pathOfDirectory)).toMatchSnapshot();
-    }, 60_000);
-
     it("init docs", async () => {
         const pathOfDirectory = await init();