+++ seed/pydantic/exhaustive/pydantic-v2/src/seed/exhaustive/resources/types/resources/object/nested_object_with_optional_field.py
@@ -23,4 +23,5 @@ class NestedObjectWithOptionalField(pydantic.BaseModel):
 
     class Config:
         allow_population_by_field_name = True
+        extra = pydantic.Extra.allow
         json_encoders = {dt.datetime: serialize_datetime}