+++ seed/pydantic/objects-with-imports/src/seed/objects_with_imports/tree.py
@@ -50,4 +50,5 @@ class Tree(pydantic.BaseModel):
         return super().dict(**kwargs_with_defaults)
 
     class Config:
+        extra = pydantic.Extra.allow
         json_encoders = {dt.datetime: serialize_datetime}