+++ seed/pydantic/trace/src/seed/trace/resources/playlist/update_playlist_request.py
@@ -28,4 +28,5 @@ class UpdatePlaylistRequest(pydantic.BaseModel):
         return super().dict(**kwargs_with_defaults)
 
     class Config:
+        extra = pydantic.Extra.allow
         json_encoders = {dt.datetime: serialize_datetime}