+++ seed/pydantic/trace/src/seed/trace/resources/submission/test_case_result_with_stdout.py
@@ -25,4 +25,5 @@ class TestCaseResultWithStdout(pydantic.BaseModel):
         return super().dict(**kwargs_with_defaults)
 
     class Config:
+        extra = pydantic.Extra.allow
         json_encoders = {dt.datetime: serialize_datetime}