+++ seed/pydantic/trace/src/seed/trace/resources/submission/workspace_files.py
@@ -26,4 +26,5 @@ class WorkspaceFiles(pydantic.BaseModel):
 
     class Config:
         allow_population_by_field_name = True
+        extra = pydantic.Extra.allow
         json_encoders = {dt.datetime: serialize_datetime}