+++ seed/pydantic/trace/src/seed/trace/resources/submission/workspace_submission_status_v_2.py
@@ -24,4 +24,5 @@ class WorkspaceSubmissionStatusV2(pydantic.BaseModel):
         return super().dict(**kwargs_with_defaults)
 
     class Config:
+        extra = pydantic.Extra.allow
         json_encoders = {dt.datetime: serialize_datetime}