+++ seed/python-sdk/examples/no-custom-config/src/seed/resources/commons/resources/types/types/metadata.py
@@ -38,4 +38,5 @@ class Metadata(pydantic.BaseModel):
         frozen = True
         smart_union = True
         allow_population_by_field_name = True
+        extra = pydantic.Extra.allow
         json_encoders = {dt.datetime: serialize_datetime}