+++ seed/python-sdk/examples/no-custom-config/src/seed/resources/types/types/movie.py
@@ -53,4 +53,5 @@ class Movie(pydantic.BaseModel):
         frozen = True
         smart_union = True
         allow_population_by_field_name = True
+        extra = pydantic.Extra.allow
         json_encoders = {dt.datetime: serialize_datetime}