+++ seed/python-sdk/exhaustive/pydantic-v1/src/seed/resources/types/resources/object/types/nested_object_with_optional_field.py
@@ -25,4 +25,5 @@ class NestedObjectWithOptionalField(pydantic.BaseModel):
         frozen = True
         smart_union = True
         allow_population_by_field_name = True
+        extra = pydantic.Extra.allow
         json_encoders = {dt.datetime: serialize_datetime}