+++ seed/python-sdk/trace/src/seed/resources/commons/types/key_value_pair.py
@@ -28,6 +28,7 @@ class KeyValuePair(pydantic.BaseModel):
     class Config:
         frozen = True
         smart_union = True
+        extra = pydantic.Extra.allow
         json_encoders = {dt.datetime: serialize_datetime}