+++ seed/python-sdk/trace/src/seed/resources/v_2/resources/v_3/resources/problem/types/get_generated_test_case_file_request.py
@@ -29,4 +29,5 @@ class GetGeneratedTestCaseFileRequest(pydantic.BaseModel):
         frozen = True
         smart_union = True
         allow_population_by_field_name = True
+        extra = pydantic.Extra.allow
         json_encoders = {dt.datetime: serialize_datetime}